Hotfix: item and order duplication merged
