+++ backend/src/main/java/shop/ninescent/mall/category/repository/GetItemListRepository.java
@@ -1,13 +1,13 @@
 package shop.ninescent.mall.category.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.ninescent.mall.order.domain.Item;
+import shop.ninescent.mall.item.domain.Item;
 
 import java.util.List;
 
-public interface GetItemListRepository extends JpaRepository<Item, Long> {
+public interface GetItemListRepository extends JpaRepository<shop.ninescent.mall.item.domain.Item, Long> {
 
-    List<Item> findBySubCategoryId(Long subCategoryId);
+    List<shop.ninescent.mall.item.domain.Item> findBySubCategoryId(Long subCategoryId);
 
     List<Item> findByItemId(Long itemId);
 }
