+++ backend/src/main/java/shop/ninescent/mall/item/repository/ItemRepository.java
@@ -2,8 +2,9 @@ package shop.ninescent.mall.item.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import shop.ninescent.mall.item.entity.Item;
+import shop.ninescent.mall.item.domain.Item;
 
 @Repository
-public interface ItemRepository extends JpaRepository<Item, Integer> {
+public interface ItemRepository extends JpaRepository<Item, Long> {
+
 }
