--- backend/src/main/java/shop/ninescent/mall/order/repository/ItemRepository.java
@@ -1,8 +1,0 @@
-package shop.ninescent.mall.order.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import shop.ninescent.mall.order.domain.Item;
-
-
-public interface ItemRepository extends JpaRepository<Item, Long> {
-}
